Rename counterfactual style files to confirm with *.styles.ts (#1338)
